requirements.txt
solaris-api>=0.0.5
pandas
pytrends
yfinance